Added integration tests for injections.
Added a new bundle (jsr330activator.testbundle3) containing three
providers for three different services, with different injections of
HelloService and HelloService2.

Created a new integration test that starts the test bundles in the
order:
 - jsr330acticator.testbundle1
 - jsr330activator.testbundle3
 - jsr330activator.testbundle2

This tests injection both with the dependent services already
present for HelloService, and not yet started (injected via OSGi events)
for HelloService2.