Simplification: catch Exception instead of individual checked exception types.
The code in ProviderAdapter.registerService() just ignores the
exceptions and ends up being a no-op.

However the exceptions are unlikely to happen at this stage: they are
reflection related, and the only way to trigger
eg. IllegalAccessException or NoSuchMethodException is by having the
provider not implement the Provider interface (and if so, the code
looking for providers wouldn't have found the class).