Embed javax.inject and Jsr330Activator in jsr330activator.testbundle3
Acomplished by mentioning the JARs that are to be embedded directly in
the dependecies list of the jsr330activator.testbundle3 POM, and then
use an Embed-Dependency directive to include them.

This POM already contained an Embed-Dependency setting for the
Jsr330Activator, but this wasn't sufficient to make maven-bundle-plugin
embed the Jsr330Activator (it needed to be listed as a direct dependency
as well).