Removed maven runtime dependency on org.apache.servicemix.bundles.javax-inject
Removed the dependency from the top POM, and put it in the POMs of the
project that needed it for compilation, and made it be a "provided"
dependency.

Without this change maven-based runtimes (e.g. Apache Felix) would
insist on loading the org.apache.servicemix.bundles.javax-inject even
when the necessary classes had been embedded and there was no OSGi
dependency for the packages.