Successfully embedded javax.inject in testbundle4, 5, 6 and 7.
What was needed was to mention the artifactId of the maven dependency
used (the bundle version of javax.inject).

Also switched the scope form "provided" to "compile" in an attempt at
losing the runtime dependency.  However this was not successful wrt to
the gogo shell provision, because that still loads the bundle version of
javax.inject even when only explicitly depending on testbundle4, 5, 6, 7
and 8 in the provisioning POM.