Avoid getting a transitive dependency on the embedded Jsr330Activator: use test scope
When using the compile scope, the resulting bundle will get a maven
dependency that will pull in the Jsr330Activator as a bundle dependency
when using maven to provision the OSGi runtime (e.g. when using gogo
shell).

The OSGi runtime availability of the Jsr330Activator does no harm, but
it isn't needed when the Jsr330Activator has been embedded and therefore
shouldn't be there.