Made it possible to access @Named annotation on injected StorageService instances.
Used the bundle version of javax.inject.  The annotations on the
injected services were dynamic proxies and there didn't seem to be any
way to access the annotation than to import the same javax.inject as the
bundles used and then use instanceof and cast to access the annotation.

Didn't bother to modify the Embed-Dependency setting since the
artifactId of the bundled version of javax.inject isn't the same as the
non-bundle javax.inject, so the bundle version won't be embedded.