Added a runtime retained @Optional annotation, allowed on fields and methods.
This annotation is used to mark injected fields and methods on a
Provider<> as optional wrt. registering the provided service with
OSGi (the normal behaviour for Jsr330Activator, is to register the
provided service with OSGi when all of the injections are satisfied, and
to retract the service from OSGi once an injection is removed).